+++ tools/HybridTester.cc
@@ -926,8 +926,8 @@ void HybridTester::SetTestGroup(BeBoard* pBoard, uint8_t pTestGroup)
     {
         for (auto cCbc : cFe->fChipVector)
         {
-            std::vector<std::pair<std::string, uint8_t>> cRegVec;
-            uint8_t cRegValue = this->to_reg ( 0, pTestGroup );
+            std::vector<std::pair<std::string, uint16_t>> cRegVec;
+            uint16_t cRegValue = this->to_reg ( 0, pTestGroup );
 
             if (cCbc->getFrontEndType() == FrontEndType::CBC3)
             {
